+++ src/redux/root.ts
@@ -3,7 +3,7 @@ import { authReducer } from "src/features/auth/auth.slice";
 import { userReducer } from "src/redux/user.slice";
 import { boardReducer } from "src/features/board/board.slice";
 import { columnReducer } from "src/redux/column.slice";
-import { cardReducer } from "src/redux/card.slice";
+import { cardReducer } from "src/features/card/card.slice";
 
 export const rootReducer = combineReducers({
   authState: authReducer,
